+++ readme/CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Version 0.5b [2016]
+* Added: Check if the user-defined os name, is different than the one identified by heuristics.
+* Added: New option "--os" that forces a user-defined os name.
 * Added: Support for testing custom HTTP headers (via "--headers" parameter).
 
 ## Version 0.4.1b [2016]